Remove unused PAL functions (dotnet/coreclr#26277)
Commit migrated from https://github.com/dotnet/coreclr/commit/5068bb8820ebd1820e70aea091bf56b19ae84f13

+++ src/coreclr/src/pal/tests/palsuite/threading/WaitForMultipleObjectsEx/test5/test5.cpp
@@ -479,7 +479,7 @@ int __cdecl main(int argc, char *argv[])
             MaxNumHelperProcess = MAX_HELPER_PROCESS;
             break;
         case 2:
-            MaxNumHelperProcess = atol(argv[1]);
+            MaxNumHelperProcess = atoi(argv[1]);
             break;
         default:
             Fail("Invalid number of arguments\n");
